Improve PauseSpec to check that the call to Browser.pause() not only completes but also does not fail. Fix conversion to milliseconds in Wait not to overflow for large int timeouts causing Browser.pause() to fail and return instantly.
Fixes geb/issues#605.

+++ module/geb-waiting/src/main/groovy/geb/waiting/Wait.groovy
@@ -127,8 +127,8 @@ class Wait {
         pass as T
     }
 
-    private static int toMilliseconds(Number seconds) {
-        seconds * 1000
+    private static long toMilliseconds(Number seconds) {
+        seconds * 1000L
     }
 
     private Instant timeoutThresholdFromNow() {
